Use UByte everywhere, keep show(Byte) via FFI

+++ effekt/shared/src/main/scala/effekt/source/Tree.scala
@@ -3,7 +3,9 @@ package source
 
 import effekt.context.Context
 import effekt.symbols.Symbol
+import effekt.util.UByte
 import kiama.util.{Source, StringSource}
+
 import scala.annotation.tailrec
 
 /**
@@ -569,7 +571,7 @@ export Term.*
 // Smart Constructors for literals
 // -------------------------------
 def UnitLit(span: Span): Literal = Literal((), symbols.builtins.TUnit, span)
-def ByteLit(value: Int, span: Span) : Literal = Literal(value, symbols.builtins.TByte, span) // assuming 0-255; `Byte` in Scala is signed
+def ByteLit(value: UByte, span: Span) : Literal = Literal(value, symbols.builtins.TByte, span)
 def IntLit(value: Long, span: Span): Literal = Literal(value, symbols.builtins.TInt, span)
 def BooleanLit(value: Boolean, span: Span): Literal = Literal(value, symbols.builtins.TBoolean, span)
 def DoubleLit(value: Double, span: Span): Literal = Literal(value, symbols.builtins.TDouble, span)
